--- consensus/state_processing/src/per_epoch_processing/gloas.rs
@@ -1,3 +1,0 @@
-pub use process_builder_pending_payments::process_builder_pending_payments;
-
-mod process_builder_pending_payments;
